refactor(routes,timer_page): rename timer_screen to timer_page
Why does Flutter not standardise the terminology... says routes -> pages or screens ...
